room questions hook
