+++ src/styles/room.scss
@@ -94,5 +94,9 @@
         }
       }
     }
+
+    .question-list {
+      margin-top: 32px;
+    }
   }
 }
